+++ addons/hr_attendance/models/__init__.py
@@ -3,6 +3,7 @@
 from . import res_config_settings
 from . import hr_attendance
 from . import hr_attendance_overtime
+from . import hr_employee_base
 from . import hr_employee
 from . import hr_employee_public
 from . import ir_ui_menu
