+++ addons/hr_attendance/models/hr_employee.py
@@ -47,24 +47,6 @@ class HrEmployee(models.Model):
             else:
                 employee.total_overtime = 0
 
-    @api.depends('user_id.im_status', 'attendance_state')
-    def _compute_presence_state(self):
-        """
-        Override to include checkin/checkout in the presence state
-        Attendance has the second highest priority after login
-        """
-        super()._compute_presence_state()
-        employees = self.filtered(lambda e: e.hr_presence_state != 'present')
-        employee_to_check_working = self.filtered(lambda e: e.attendance_state == 'checked_out'
-                                                            and e.hr_presence_state == 'to_define')
-        working_now_list = employee_to_check_working._get_employee_working_now()
-        for employee in employees:
-            if employee.attendance_state == 'checked_out' and employee.hr_presence_state == 'to_define' and \
-                    employee.id not in working_now_list:
-                employee.hr_presence_state = 'absent'
-            elif employee.attendance_state == 'checked_in':
-                employee.hr_presence_state = 'present'
-
     def _compute_hours_last_month(self):
         now = fields.Datetime.now()
         now_utc = pytz.utc.localize(now)
@@ -196,9 +178,3 @@ class HrEmployee(models.Model):
             raise exceptions.UserError(_('Cannot perform check out on %(empl_name)s, could not find corresponding check in. '
                 'Your attendances have probably been modified manually by human resources.') % {'empl_name': self.sudo().name, })
         return attendance
-
-    def _compute_presence_icon(self):
-        res = super()._compute_presence_icon()
-        # All employee must chek in or check out. Everybody must have an icon
-        self.filtered(lambda employee: not employee.show_hr_icon_display).show_hr_icon_display = True
-        return res
