+++ addons/hr_holidays/models/__init__.py
@@ -2,6 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import resource
+from . import hr_employee_base
 from . import hr_employee
 from . import hr_department
 from . import hr_leave
