+++ addons/hr_presence/models/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import hr_employee_base
 from . import hr_employee
 from . import ir_websocket
 from . import res_company
